reformat code

+++ src/main/java/com/org/ita/kata/implementation/nastiakomarenko/SIxImpl.java
@@ -9,7 +9,7 @@ public class SIxImpl implements Six {
         long m = 0;
         long i = 0;
         while (m < M) {
-            m += i*i*i;
+            m += i * i * i;
             if (m == M) {
                 return i;
             }
@@ -26,15 +26,15 @@ public class SIxImpl implements Six {
 
     @Override
     public double f(double x) {
-         x =(Math.exp(1)) - 15;
-        double answer =0;
-        if(x<0){
-            double mult = x *(-1);
-             answer = (double) Math.sqrt(1 + mult) - 1;
+        x = (Math.exp(1)) - 15;
+        double answer = 0;
+        if (x < 0) {
+            double mult = x * (-1);
+            answer = Math.sqrt(1 + mult) - 1;
 
             System.out.println(answer);
-        }else {
-             answer = (double) Math.sqrt(1 + x) - 1;
+        } else {
+            answer = Math.sqrt(1 + x) - 1;
         }
 
         return answer;
